mergeConfig do not need to return the config since we pass the pointer
